Remove decimal formatting from ops/sec metrics
- Change Reads, Writes, and Total ops display from %.1f to %d
- Display whole numbers instead of decimals (e.g., 1613/sec instead of 1613.2/sec)
- Improves readability of metrics output

+++ src/main/java/com/redis/demo/metrics/MetricsCollector.java
@@ -200,10 +200,10 @@ public class MetricsCollector implements Runnable {
             long writesDelta = currentWrites - lastWriteCount;
 
             System.out.println();
-            System.out.printf("Reads:       %.1f/sec%n", (double) readsDelta / intervalSeconds);
-            System.out.printf("Writes:      %.1f/sec%n", (double) writesDelta / intervalSeconds);
-            System.out.printf("Total ops:   %.1f/sec%n",
-                            (double) (readsDelta + writesDelta) / intervalSeconds);
+            System.out.printf("Reads:       %d/sec%n", readsDelta / intervalSeconds);
+            System.out.printf("Writes:      %d/sec%n", writesDelta / intervalSeconds);
+            System.out.printf("Total ops:   %d/sec%n",
+                            (readsDelta + writesDelta) / intervalSeconds);
 
             // Update last counts for next interval
             lastReadCount = currentReads;
